% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_output_helpers.R
\name{make_palette}
\alias{make_palette}
\title{make_palette
Because the dummy data contain less data points we
need to provided a shorter named palette}
\usage{
make_palette()
}
\description{
make_palette
Because the dummy data contain less data points we
need to provided a shorter named palette
}
