% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{prep_cohorts_data}
\alias{prep_cohorts_data}
\title{prep_cohorts_data
Function to prepare the cohorts dataset for plotting}
\usage{
prep_cohorts_data(dat, geo_group)
}
\arguments{
\item{dat}{}

\item{geo_group}{}
}
\value{
a formatted dataframe
}
\description{
prep_cohorts_data
Function to prepare the cohorts dataset for plotting
}
