% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_input_helpers.R
\name{render_input_for_cohorts_table}
\alias{render_input_for_cohorts_table}
\title{render_input_for_cohorts_table}
\usage{
render_input_for_cohorts_table(ns, dat, geo_group)
}
\arguments{
\item{ns}{namespace function}

\item{dat}{dataframe}

\item{geo_group}{"area", "county" or "country"}
}
\value{
a taglist to be rendered in the UI
for the table
}
\description{
render_input_for_cohorts_table
}
