% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_input_helpers.R
\name{render_input_for_mortality_month_plot}
\alias{render_input_for_mortality_month_plot}
\title{render_input_for_mortality_month_plot}
\usage{
render_input_for_mortality_month_plot(ns, dat, viz)
}
\arguments{
\item{ns}{namespace function}

\item{dat}{dataframe}

\item{viz}{"zone", "fylke" or "all"}
}
\value{
a taglist to be rendered in the UI
for the plot
}
\description{
render_input_for_mortality_month_plot
}
