% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_input_helpers.R
\name{render_input_for_plot}
\alias{render_input_for_plot}
\title{render_input_for_plot}
\usage{
render_input_for_plot(ns, dat, viz)
}
\arguments{
\item{ns}{}

\item{dat}{}

\item{viz}{}
}
\value{
a taglist to be rendered in the UI
for the plot
}
\description{
render_input_for_plot
}
