% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_input_helpers.R
\name{render_input_for_table}
\alias{render_input_for_table}
\title{render_input_for_table}
\usage{
render_input_for_table(ns, dat, viz)
}
\arguments{
\item{viz}{}
}
\value{
a taglist to be rendered in the UI
for the table
}
\description{
render_input_for_table
}
