% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ui_helpers.R
\name{select_months}
\alias{select_months}
\title{select_months}
\usage{
select_months(id, digit = F, multiple = T)
}
\arguments{
\item{id}{id for the selectInput}

\item{digit}{TRUE/FALSE for month as digits or text}

\item{multiple}{TRUE/FALSE for multiple selection}
}
\value{
a vector to be used in selectInput
}
\description{
Function to return a vector of months to be rendered in the UI.
}
