% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{select_year}
\alias{select_year}
\title{selesc_year}
\usage{
select_year(id, multiple = F, resolution = "m")
}
\arguments{
\item{id}{id for the selectInput}

\item{multiple}{TRUE/FALSE for multiple selection}

\item{resolution}{"m" for monthly data, "y" for yearly data}
}
\value{
a vector to be used in selectInput
}
\description{
Function to return a vector of years to be rendered in the UI.
Years are hardcoded for now.
Monthly data resolution has data from the current year,
yearly data resolution has data up to last full year.
}
