% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{prep_cohorts_data}
\alias{prep_cohorts_data}
\title{prep_cohorts_data}
\usage{
prep_cohorts_data(dat)
}
\arguments{
\item{dat}{a data frame}
}
\value{
a formatted dataframe
}
\description{
Function to prepare the cohorts dataset for plotting
}
